Cleaning code that is not needed
Removing stuff from OfflineConnectors. Setting offline version of
application connection to be used in offline mode to avoid NPE
when accessing layoutmanager. float/int/... encoding still broken.
Not sure if because of 7.2.4 bug, or if those will be actually
always decoded as strings on client side. If so -> have to write
custom encoder or decoder.

Fixes #4